Fix clang-tidy readability-redundant-casting in render_terrain
TransformState::getZoom() already returns double, so the static_cast<double>
around it is redundant and fails the Linux clang-tidy build (-Werror).

+++ src/mbgl/renderer/render_terrain.cpp
@@ -1192,7 +1192,7 @@ void RenderTerrain::updateInstancedDepthUBO(PaintParameters& parameters) {
     }
 
     // Shared evaluated props (unpack / exaggeration / skirt offset), same as the tweaker.
-    const auto zoom = std::max(static_cast<double>(parameters.state.getZoom()), 0.0);
+    const auto zoom = std::max(parameters.state.getZoom(), 0.0);
     const float elevationOffset = static_cast<float>(util::M2PI * util::EARTH_RADIUS_M / std::pow(2.0, zoom) / 5.0);
     const shaders::TerrainEvaluatedPropsUBO propsUBO = {.unpack = getDEMUnpackVector(),
                                                         .exaggeration = getExaggeration(),
